requirements.txt
pip>=25.1.1
flask==2.0.1
pyodbc==4.0.32
qrcode[pil]==7.3.1
pillow==8.4.0
flask-login==0.5.0
flask-wtf==0.15.1
python-dotenv==0.19.0